Added FractalNoise
Added a new type of noise: fractal noise, which is based on layers (or octaves) of noise being added on top of each other.
Also changed the RNG, yet again, to a hashing algorithm that generates numbers based on coordinates, which is required for fractal noise.
I've provided an example page for the noise as well.