Added GradientMap
Added the gradient map filter which assigns colors from a gradient based on the source value.
So, if the source pixel has a value of 0.2, it gets set to the color of the gradient at position 0.2.

Works best with grayscale "height maps."

As a side effect, color interpolators accept positions beyond 0 and 1 now.